fix(neck): harden playback continuity and split Neck into partials
Keep silence until TrackStatus Running, ignore premature Ended, make pause
idempotent, prefetch on enqueue, and split Neck by Struna/playback/queue/FIFO.
